+++ kyma/utils/__init__.py
@@ -1,1 +1,12 @@
 """Shared utility helpers."""
+
+from kyma.utils.env import loadrepowandbenv, parseenv
+from kyma.utils.wandb import WandbRun, createwandbrun, haswandbnetrc
+
+__all__ = [
+    "WandbRun",
+    "createwandbrun",
+    "haswandbnetrc",
+    "loadrepowandbenv",
+    "parseenv",
+]